requirements.txt
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
drf-yasg==1.21.4